Rename TRIPS processor
